refactor: create component folders for Dialog & Divider
