+++ src/components/Dialog/Dialog.story.vue
@@ -1,7 +1,7 @@
 <script setup lang="ts">
 import { ref } from 'vue'
 import Dialog from './Dialog.vue'
-import { Button } from './Button'
+import { Button } from '../Button'
 
 const dialog1 = ref(false)
 const dialog2 = ref(false)
