+++ src/components/Dialog/index.ts
@@ -1,0 +1,2 @@
+export { default as Dialog } from './Dialog.vue'
+export type { DialogProps } from './types'
