+++ src/components/Divider/index.ts
@@ -1,0 +1,2 @@
+export { default as Divider} from './Divider.vue'
+export type { DividerProps } from './types'
