+++ src/components/Divider/types.ts
@@ -1,0 +1,12 @@
+interface DividerAction {
+  label: string
+  handler: () => any
+  loading?: boolean
+}
+
+export interface DividerProps {
+  orientation?: 'horizontal' | 'vertical'
+  position?: 'start' | 'center' | 'end'
+  flexItem?: boolean
+  action?: DividerAction
+}
