+++ src/index.ts
@@ -11,9 +11,9 @@ export * from './components/Checkbox'
 export { default as DatePicker } from './components/DatePicker/DatePicker.vue'
 export { default as DateTimePicker } from './components/DatePicker/DateTimePicker.vue'
 export { default as DateRangePicker } from './components/DatePicker/DateRangePicker.vue'
-export { default as Dialog } from './components/Dialog.vue'
+export * from './components/Dialog'
 export { default as Dialogs } from './components/Dialogs.vue'
-export { default as Divider } from './components/Divider.vue'
+export * from './components/Divider'
 export * from './components/Dropdown'
 export * from './components/ErrorMessage'
 export { default as FeatherIcon } from './components/FeatherIcon.vue'
